# Conflicts:
#	src/main/java/models/BoardsEntity.java
#	src/main/java/models/UsersEntity.java
#	src/main/java/services/BoardService.java
#	src/main/java/services/UserService.java

+++ src/main/java/utils/VotesHibernateUtil.java
@@ -12,7 +12,7 @@ public class VotesHibernateUtil {
     /**
      * Read all boards of the user.
      */
-    public static VotesEntity saveVote(long option_id, long user_id, int val) {
+    public static VotesEntity saveVote(long option_id, long user_id, long val) {
         VotesEntity vote;
         EntityManager manager = ENTITY_MANAGER_FACTORY.createEntityManager();
         EntityTransaction transaction = null;
@@ -47,7 +47,7 @@ public class VotesHibernateUtil {
     /**
      * Read all boards of the user.
      */
-    public static VotesEntity editVote(long vote_id, long user_id, int val) {
+    public static VotesEntity editVote(long vote_id, long user_id, long val) {
         VotesEntity vote;
         EntityManager manager = ENTITY_MANAGER_FACTORY.createEntityManager();
         EntityTransaction transaction = null;
